major: auth.js aufgeteilt in login.js & create_account.js
