+++ public/html/login.html
@@ -46,6 +46,6 @@
           <p class="register-link">Noch keinen Account? <a href="register.html">Registriere dich hier</a></p>
       </div>
   </div>
-   <script type="module" src="../js/auth.js"></script>
+   <script type="module" src="../js/login.js"></script>
   </body>
 </html>
