+++ public/html/register.html
@@ -55,6 +55,6 @@
             <p class="login-link">Bereits einen Account? <a href="login.html">Logge dich hier ein</a></p>
         </div>
     </div>
-    <script type="module" src="../js/auth.js"></script>
+    <script type="module" src="../js/create_account.js"></script>
 </body>
 </html>
